accessible/tests/mochitest/relations/test_update.html

<html>

<head>
  <title>Test updating of accessible relations</title>
  <link rel="stylesheet" type="text/css"
        href="chrome://mochikit/content/tests/SimpleTest/test.css" />

  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <script type="application/javascript"
          src="../common.js"></script>
  <script type="application/javascript"
          src="../relations.js"></script>
  <script type="application/javascript"
          src="../role.js"></script>
  <script type="application/javascript"
          src="../events.js"></script>

  <script type="application/javascript">
    function testRelated(aRelAttr, aHostRelation, aDependentRelation,
                         aHostID, aHostNodeID, aDependent1ID, aDependent2ID) {
      // no attribute
      testRelation(aDependent1ID, aDependentRelation, null);
      testRelation(aDependent2ID, aDependentRelation, null);
      if (aHostRelation)
        testRelation(aHostID, aHostRelation, null);

      // set attribute
      getNode(aHostNodeID).setAttribute(aRelAttr, aDependent1ID);
      testRelation(aDependent1ID, aDependentRelation, aHostID);
      testRelation(aDependent2ID, aDependentRelation, null);
      if (aHostRelation)
        testRelation(aHostID, aHostRelation, aDependent1ID);

      // change attribute
      getNode(aHostNodeID).setAttribute(aRelAttr, aDependent2ID);
      testRelation(aDependent1ID, aDependentRelation, null);
      testRelation(aDependent2ID, aDependentRelation, aHostID);
      if (aHostRelation)
        testRelation(aHostID, aHostRelation, aDependent2ID);

      // remove attribute
      getNode(aHostNodeID).removeAttribute(aRelAttr);
      testRelation(aDependent1ID, aDependentRelation, null);
      testRelation(aDependent2ID, aDependentRelation, null);
      if (aHostRelation)
        testRelation(aHostID, aHostRelation, null);
    }

    async function insertRelated(aHostRelAttr, aDependentID, aInsertHostFirst,
                                 aHostRelation, aDependentRelation) {
      let hostNode;
      let dependentNode;
      let reordered = waitForEvent(EVENT_REORDER, document);

      hostNode = document.createElement("div");
      hostNode.setAttribute(aHostRelAttr, aDependentID);

      dependentNode = document.createElement("div");
      dependentNode.setAttribute("id", aDependentID);

      if (aInsertHostFirst) {
        document.body.appendChild(hostNode);
        document.body.appendChild(dependentNode);
      } else {
        document.body.appendChild(dependentNode);
        document.body.appendChild(hostNode);
      }

      await reordered;

      testRelation(dependentNode, aDependentRelation, hostNode);
      if (aHostRelation)
        testRelation(hostNode, aHostRelation, dependentNode);
    }

    /**
     * Relative accessible recreation shouldn't break accessible relations.
     * Note: modify this case if the invoke function doesn't change accessible
     * tree due to changes in layout module. It can be changed on any case
     * when accessibles are recreated.
     */
    async function recreateRelatives(aContainerID, aLabelID, aElmID) {
      const containerNode = getNode(aContainerID);
      const container = getAccessible(containerNode);

      testRelation(aLabelID, RELATION_LABEL_FOR, aElmID);
      testRelation(aElmID, RELATION_LABELLED_BY, aLabelID);

      let events = waitForEvents([
        [EVENT_HIDE, container],
        [EVENT_SHOW, containerNode],
      ]);
      containerNode.setAttribute('role', 'group');
      await events;

      testRelation(aLabelID, RELATION_LABEL_FOR, aElmID);
      testRelation(aElmID, RELATION_LABELLED_BY, aLabelID);
    }

    async function doTest() {
      // Relation updates on ARIA attribute changes.
      testRelated("aria-labelledby",
                  RELATION_LABELLED_BY, RELATION_LABEL_FOR,
                  "host", "host", "dependent1", "dependent2");

      testRelated("aria-describedby",
                  RELATION_DESCRIBED_BY, RELATION_DESCRIPTION_FOR,
                  "host", "host", "dependent1", "dependent2");

      testRelated("aria-controls",
                  RELATION_CONTROLLER_FOR, RELATION_CONTROLLED_BY,
                  "host", "host", "dependent1", "dependent2");

      testRelated("aria-flowto",
                  RELATION_FLOWS_TO, RELATION_FLOWS_FROM,
                  "host", "host", "dependent1", "dependent2");

      // Document relation updates on ARIA attribute change.
      testRelated("aria-labelledby",
                  RELATION_LABELLED_BY, RELATION_LABEL_FOR,
                  document, "body", "dependent1", "dependent2");

      // Insert related accessibles into tree.
      await insertRelated("aria-labelledby", "dependent3", true,
                          RELATION_LABELLED_BY, RELATION_LABEL_FOR);
      await insertRelated("aria-labelledby", "dependent4", false,
                          RELATION_LABELLED_BY, RELATION_LABEL_FOR);

      await insertRelated("aria-describedby", "dependent5", true,
                          RELATION_DESCRIBED_BY,
                          RELATION_DESCRIPTION_FOR);
      await insertRelated("aria-describedby", "dependent6", false,
                          RELATION_DESCRIBED_BY,
                          RELATION_DESCRIPTION_FOR);

      await insertRelated("aria-controls", "dependent9", true,
                          RELATION_CONTROLLER_FOR,
                          RELATION_CONTROLLED_BY);
      await insertRelated("aria-controls", "dependent10", false,
                          RELATION_CONTROLLER_FOR,
                          RELATION_CONTROLLED_BY);

      await insertRelated("aria-flowto", "dependent11", true,
                          RELATION_FLOWS_TO, RELATION_FLOWS_FROM);
      await insertRelated("aria-flowto", "dependent12", false,
                          RELATION_FLOWS_TO, RELATION_FLOWS_FROM);

      // Update relations when accessibles are recreated
      await recreateRelatives("container", "label", "input");

      SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();
    addA11yLoadEvent(doTest);
  </script>

</head>

<body id="body">

  <a target="_blank"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=573469"
     title="Cache relations defined by ARIA attributes">
    Mozilla Bug 573469
  </a>
  <a target="_blank"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=631068"
     title="Accessible recreation breaks relations">
    Mozilla Bug 631068
  </a>
  <a target="_blank"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=635346"
     title="Allow relations for document defined on document content">
    Mozilla Bug 635346
  </a>
  <br>
  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test">
  </pre>

  <div id="dependent1">label</div>
  <div id="dependent2">label2</div>
  <div role="checkbox" id="host"></div>

  <form id="container" style="overflow: hidden;">
    <label for="input" id="label">label</label>
    <input id="input">
  </form>
</body>
</html>
